migrating names, libxml needs to stabilize with thread-safety or test suite is flaky
