+++ Cargo.toml
@@ -7,8 +7,8 @@ authors = ["Jan Frederik Schaefer <j.schaefer@jacobs-university.de>"]
 name = "llamapun"
 crate-type = ["rlib", "dylib", "staticlib"]
 
-[dependencies.rustlibxml]
-git = "https://github.com/jfschaefer/rust-libxml.git"
+[dependencies.libxml]
+git = "https://github.com/dginev/rust-libxml.git"
 [dependencies.unidecode]
 git = "https://github.com/chowdhurya/rust-unidecode.git"
 [dependencies.rustmorpha]
