+++ README.md
@@ -1,4 +1,4 @@
 [![Build Status](https://secure.travis-ci.org/dginev/rust-llamapun.png?branch=master)](http://travis-ci.org/dginev/rust-llamapun)
 
 # rust-llamapun
-Trying to reimplement the kwarc/LLaMaPUn project in Rust
+Port of the KWARC/LLaMaPUn project in Rust
